requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-dotenv==1.1.0
pydantic==2.11.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-multipart==0.0.20
email-validator==2.3.0
